+++ src/main/java/repast/model/agrosuccess/params/EnvrModelParams.java
@@ -1,5 +1,6 @@
 package repast.model.agrosuccess.params;
 
+import me.ajlane.geo.repast.fire.FireParams;
 import me.ajlane.geo.repast.seeddispersal.SeedDispersalParams;
 import me.ajlane.geo.repast.seeddispersal.SeedViabilityParams;
 import me.ajlane.geo.repast.soilmoisture.SoilMoistureParams;
@@ -27,5 +28,10 @@ public interface EnvrModelParams {
    *         discrete states.
    */
   public SoilMoistureParams getSoilMoistureParams();
+
+  /**
+   * @return Parameters needed to run the fire ignition and spread model.
+   */
+  public FireParams getFireParams();
 
 }